remove file layers.py
